requirements.txt
Flask==2.0.2
mysql-connector-python
werkzeug==2.0.2
